backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
openai==1.51.0
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.10.3
httpx==0.25.2
sqlalchemy==2.0.23
python-dateutil==2.8.2
click==8.1.7
rich==13.7.0
firebase-functions==0.1.0
firebase-admin==6.2.0 